+++ insts/std/METAFLUTE/LSFLUTE.cpp
@@ -23,7 +23,7 @@ LSFLUTE::LSFLUTE() : Instrument()
 	// future setup here?
 }
 
-int LSFLUTE::init(float p[], int n_args)
+int LSFLUTE::init(double p[], int n_args)
 {
 // p0 = start; p1 = dur; p2 = noise amp; p3 = length1; p4 = length2
 // p5 = amp multiplier; p6 = stereo spread (0-1) <optional>
