+++ insts/std/MMOVE/MBASE.cpp
@@ -71,7 +71,7 @@ MBASE::~MBASE()
    }
 }
 
-int MBASE::init(float p[], int n_args)
+int MBASE::init(double p[], int n_args)
 {
     int    flag, UseMikes, cartflag = 0;
     float  outskip, inskip, abs_factor, rvb_time;
