+++ insts/std/STEREO/STEREO.cpp
@@ -33,7 +33,7 @@ STEREO::~STEREO()
 
 #define MATRIX_PFIELD_OFFSET 4
 
-int STEREO::init(float p[], int n_args)
+int STEREO::init(double p[], int n_args)
 {
 // p0 = outsk; p1 = insk; p2 = dur (-endtime); p3 = amp; p4-n = channel mix matrix
 // we're stashing the setline info in gen table 1
