+++ insts/std/TRANS/TRANS.cpp
@@ -64,7 +64,7 @@ TRANS :: ~TRANS()
 }
 
 
-int TRANS :: init(float p[], int n_args)
+int TRANS :: init(double p[], int n_args)
 {
    float outskip, inskip, dur, transp, interval, total_indur, dur_to_read;
 	int rvin;
